+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.12.1](https://github.com/CrowdHailer/raxx/tree/0.12.1) - 2017-09-24
+
+## Fixed
+- Cannot set HTTP headers with uppercase attribute.
+- Cannot add an HTTP header twice.
+
 ## [0.12.0](https://github.com/CrowdHailer/raxx/tree/0.12.0) - 2017-08-31
 
 ## Changed